LayoutTests/mathml/presentation/msubsup-subscriptshift-expected.txt

Tests that subscriptshift works on msubsup. After adding subscriptshift the height of msubsup element should increase

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS height > 150 is true
PASS successfullyParsed is true

TEST COMPLETE
